% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_geo.R
\docType{data}
\name{country_geo}
\alias{country_geo}
\title{Country geographic information}
\format{An object of class \code{data.frame} with 142 rows and 4 columns.}
\source{
\url{https://opendata.socrata.com/dataset/Country-List-ISO-3166-Codes-Latitude-Longitude/mnkm-8ram}
        
        \url{https://www.kaggle.com/fernandol/countries-of-the-world}
}
\usage{
country_geo
}
\description{
Data frame of Gapminder country names, average latitudes and longitudes, and surface areas:
\describe{
\item{latitude}{Average latitude.}
\item{longitude}{Average longitude.}
\item{area}{Surface area in square kilometer.}
}
}
\examples{
if (require("dplyr")) {
gapminder \%>\%
  filter(year == 2007, continent == "Asia") \%>\%
  left_join(country_geo, by = "country")
}
}
\seealso{
\code{\link{gapminder}} for a description of the dataset
}
\keyword{datasets}
