% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder_unfiltered.R
\docType{data}
\name{gapminder_unfiltered}
\alias{gapminder_unfiltered}
\title{Gapminder data, unfiltered.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3313 rows and 6 columns.
}
\usage{
gapminder_unfiltered
}
\description{
The supplemental data frame \code{gapminder_unfiltered} was not filtered on
\code{year} or for complete data and has 3313 rows. Everything else is as
documented in \code{\link{gapminder}}.
}
\keyword{datasets}
